--- a/tree.go
+++ b/tree.go
@@ -118,6 +118,11 @@ type node struct {
 	fullPath  string
 }
 
+type skip struct {
+	path      string
+	paramNode *node
+}
+
 // Increments priority of the given child and reorders if necessary
 func (n *node) incrementChildPrio(pos int) int {
 	cs := n.children
@@ -400,6 +405,8 @@ type nodeValue struct {
 // made if a handle exists with an extra (without the) trailing slash for the
 // given path.
 func (n *node) getValue(path string, params *Params, unescape bool) (value nodeValue) {
+	var skipped *skip
+
 walk: // Outer loop for walking the tree
 	for {
 		prefix := n.path
@@ -411,23 +418,22 @@ walk: // Outer loop for walking the tree
 				idxc := path[0]
 				for i, c := range []byte(n.indices) {
 					if c == idxc {
-						childWillWalkTo := n.children[i]
-
-						if strings.Split(childWillWalkTo.path, "/")[0] != strings.Split(path, "/")[0] && strings.HasPrefix(n.children[len(n.children)-1].path, ":") {
-							for _, child := range childWillWalkTo.children {
-								cPath := string(c) + child.path
-
-								if cPath == path {
-									child.path = cPath
-									n = child
-									continue walk
-								}
+						if strings.HasPrefix(n.children[len(n.children)-1].path, ":") {
+							skipped = &skip{
+								path: prefix + path,
+								paramNode: &node{
+									path:      n.path,
+									wildChild: n.wildChild,
+									nType:     n.nType,
+									priority:  n.priority,
+									children:  n.children,
+									handlers:  n.handlers,
+									fullPath:  n.fullPath,
+								},
 							}
-
-							continue
 						}
 
-						n = childWillWalkTo
+						n = n.children[i]
 						continue walk
 					}
 				}
@@ -556,6 +562,13 @@ walk: // Outer loop for walking the tree
 			}
 
 			return
+		}
+
+		if path != "/" && strings.HasSuffix(skipped.path, path) {
+			path = skipped.path
+			n = skipped.paramNode
+			skipped = nil
+			continue walk
 		}
 
 		// Nothing found. We can recommend to redirect to the same URL with an