Signed-off-by: Yue Yang <g1enyy0ung@gmail.com>

Close #2682.
Close #2696.

This PR solves the above issues.

~~For a detailed description, see https://github.com/gin-gonic/gin/issues/2682#issuecomment-824579980.~~