@g1eny0ung 2021-04-28
> @g1eny0ung Can you also send the PR to https://github.com/julienschmidt/httprouter?

Because the PR https://github.com/julienschmidt/httprouter/pull/329 still not be merged, so it's needed to wait for it or let @rw-access update the related code.