@appleboy 2021-04-28
@g1eny0ung Please provide the benchmark result like https://github.com/gin-gonic/gin/pull/2663#issuecomment-811665303 and https://github.com/gin-gonic/gin/pull/2663#issuecomment-813111923