@g1eny0ung 2021-04-29
@appleboy After @rw-access noted (backtracking) and some thoughts 🤔, I update the code to save a temporary `skipped` value when the children have a param node at the last.

The skipped contains the current node path and a param node (same as the current node except no indices). When a param path has the same prefix with the exact path, the prefix tree will work as before, keep searching downwards, but it will fail. 

Then the follow-up logic will check the `skipped` value is not `nil` to perform a backtracking search. The param node in `skipped` will be set to the current node. Because it has no **indices**, it will jump the non-wildcard children check, go directly to the place of param judgment.

Compared to before, I have added more tests, if there are omissions, please help me to add. 😆

The benchmarks:

**master** - https://travis-ci.com/github/justttype/go-http-routing-benchmark/builds/224357913
**this PR** - https://travis-ci.com/github/justttype/go-http-routing-benchmark/builds/224455608

The `benchstat` report:

```
name              old time/op    new time/op    delta
Gin_Param           66.6ns ± 0%    67.6ns ± 0%   ~     (p=1.000 n=1+1)
Gin_Param5           117ns ± 0%     112ns ± 0%   ~     (p=1.000 n=1+1)
Gin_Param20          303ns ± 0%     298ns ± 0%   ~     (p=1.000 n=1+1)
Gin_ParamWrite       123ns ± 0%     119ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GithubStatic    77.4ns ± 0%    90.0ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GithubParam      133ns ± 0%     156ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GithubAll       26.8µs ± 0%    31.1µs ± 0%   ~     (p=1.000 n=1+1)
Gin_GPlusStatic     61.7ns ± 0%    66.8ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GPlusParam      82.5ns ± 0%    92.2ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GPlus2Params     103ns ± 0%     116ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GPlusAll        1.13µs ± 0%    1.29µs ± 0%   ~     (p=1.000 n=1+1)
Gin_ParseStatic     62.3ns ± 0%    65.5ns ± 0%   ~     (p=1.000 n=1+1)
Gin_ParseParam      71.4ns ± 0%    74.3ns ± 0%   ~     (p=1.000 n=1+1)
Gin_Parse2Params    84.4ns ± 0%    84.7ns ± 0%   ~     (p=1.000 n=1+1)
Gin_ParseAll        2.01µs ± 0%    2.16µs ± 0%   ~     (p=1.000 n=1+1)
Gin_StaticAll       18.5µs ± 0%    23.7µs ± 0%   ~     (p=1.000 n=1+1)

name              old alloc/op   new alloc/op   delta
Gin_Param            0.00B          0.00B        ~     (all equal)
Gin_Param5           0.00B          0.00B        ~     (all equal)
Gin_Param20          0.00B          0.00B        ~     (all equal)
Gin_ParamWrite       0.00B          0.00B        ~     (all equal)
Gin_GithubStatic     0.00B          0.00B        ~     (all equal)
Gin_GithubParam      0.00B          0.00B        ~     (all equal)
Gin_GithubAll        0.00B          0.00B        ~     (all equal)
Gin_GPlusStatic      0.00B          0.00B        ~     (all equal)
Gin_GPlusParam       0.00B          0.00B        ~     (all equal)
Gin_GPlus2Params     0.00B          0.00B        ~     (all equal)
Gin_GPlusAll         0.00B          0.00B        ~     (all equal)
Gin_ParseStatic      0.00B          0.00B        ~     (all equal)
Gin_ParseParam       0.00B          0.00B        ~     (all equal)
Gin_Parse2Params     0.00B          0.00B        ~     (all equal)
Gin_ParseAll         0.00B          0.00B        ~     (all equal)
Gin_StaticAll        0.00B          0.00B        ~     (all equal)

name              old allocs/op  new allocs/op  delta
Gin_Param             0.00           0.00        ~     (all equal)
Gin_Param5            0.00           0.00        ~     (all equal)
Gin_Param20           0.00           0.00        ~     (all equal)
Gin_ParamWrite        0.00           0.00        ~     (all equal)
Gin_GithubStatic      0.00           0.00        ~     (all equal)
Gin_GithubParam       0.00           0.00        ~     (all equal)
Gin_GithubAll         0.00           0.00        ~     (all equal)
Gin_GPlusStatic       0.00           0.00        ~     (all equal)
Gin_GPlusParam        0.00           0.00        ~     (all equal)
Gin_GPlus2Params      0.00           0.00        ~     (all equal)
Gin_GPlusAll          0.00           0.00        ~     (all equal)
Gin_ParseStatic       0.00           0.00        ~     (all equal)
Gin_ParseParam        0.00           0.00        ~     (all equal)
Gin_Parse2Params      0.00           0.00        ~     (all equal)
Gin_ParseAll          0.00           0.00        ~     (all equal)
Gin_StaticAll         0.00           0.00        ~     (all equal)
```

Thank you all for your reviews.