@daheige 2021-05-14
nice,great job!


------------------ Original ------------------
From: Denis Saponenko ***@***.***&gt;
Date: Wed,May 12,2021 7:13 PM
To: gin-gonic/gin ***@***.***&gt;
Cc: heige ***@***.***&gt;, Comment ***@***.***&gt;
Subject: Re: [gin-gonic/gin] Fix conflict between param and exact path (#2706)





 
It's very important fix for everybody!
 
—
You are receiving this because you commented.
Reply to this email directly, view it on GitHub, or unsubscribe.