@appleboy 2021-05-18
I will bump the new version v1.7.2 if the PR has been merge. Waiting for @thinkerou approval.