@appleboy 2021-05-21
Hi All,

We already released https://github.com/gin-gonic/gin/releases/tag/v1.7.2 version, just include the changes.